+++ CHANGELOG.md
@@ -59,6 +59,7 @@ These classes will run as threads in the next releases:
 
 - [matterbridge]: Add a persisted `uuid` properties to PlatformMatterbridge. This allows to detect a new setup in the plugin code.
 - [eslint]: Remove `eslint-plugin-promise` (not actively maintained) and add optional @typescript-eslint promise rules.
+- [utils]: Add fireAndForget() helper.
 
 ### Added